app/requirements.txt
Flask==2.2.4
python-consul
